pyproject.toml
[project]
name = "market-pulse"
version = "0.1.0"
description = "Market Pulse - Lean MVP for market news analytics"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "vaderSentiment>=3.3.2",
    "pyyaml>=6.0.0",
    "python-dateutil>=2.8.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "jupyter>=1.1.1",
    "ipykernel>=6.30.1",
    "pandas>=2.3.2",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
    "praw>=7.7.0",
    "prawcore>=2.4.0",
    "python-dotenv>=1.0.0",
    "transformers>=4.30.0,<4.40.0",
    "torch==2.0.1",
    "numpy>=1.24.0,<2.0.0",
    "torchvision==0.15.2",
    "yfinance>=0.2.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ['py311']
line-length = 88

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "beautifulsoup4>=4.13.5",
    "black>=25.1.0",
    "mypy>=1.18.1",
    "pytest>=8.4.2",
    "ruff>=0.13.0",
]
